Move server and client modules to net package.
