+++ bin/snakes
@@ -26,7 +26,7 @@ import sys
 import os
 import os.path
 
-from snakem.role import server
+from snakem.net import server
 
 def set_path():
     """Add snakem package to sys.path the gPodder way.
